fixup null datum value
