+++ packages/studio-base/src/components/TimeBasedChart/types.ts
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import type { ChartData } from "@foxglove/studio-base/components/Chart";
+import type { ChartData } from "@foxglove/studio-base/components/Chart/types";
 
 // alias types for convenience
 type ChartDatasets = ChartData["datasets"];
